Bug fix and refactor user initialization
-- Move intializeUser logic to new hook 'useInitializeUser' and refactor for reusability and focus.
-- Fixed infinite loop bug caused by redirect.
