+++ src/views/AuthLandingPage.jsx
@@ -1,43 +1,22 @@
-import React, { useEffect, useRef } from "react";
-import { useAuth0 } from "@auth0/auth0-react";
-import { useNavigate } from "react-router-dom";
-import User from "../models/User";
-import { useApp } from "../contexts/AppContext";
+import React, { useEffect } from "react";
 import FullPageLoader from "../components/FullPageLoader";
+import useInitializeUser from "../hooks/useInitializeUser";
+import { useApp } from "../contexts/AppContext";
+import { useNavigate } from "react-router-dom";
 
-const LoadingPage = () => {
-    const { isAuthenticated, user: auth0User } = useAuth0();
-    const { setUser } = useApp();
+const AuthLandingPage = () => {
+    const { currentUser } = useApp();
     const navigate = useNavigate();
-    const isInitialized = useRef(false); // Ref to track initialization
+
+    useInitializeUser(!currentUser, true); // Skip redirect in the hook
 
     useEffect(() => {
-        const initializeUser = async () => {
-            if (isInitialized.current) return; // Prevent re-initialization
-            if (isAuthenticated && auth0User) {
-                isInitialized.current = true; // Mark as initialized
-
-                // Transform Auth0 data into a User object
-                const currentUser = new User({
-                    userID: auth0User.sub,
-                    userName: auth0User.name,
-                    email: auth0User.email,
-                    settings: auth0User.user_metadata || {}, // Include metadata
-                });
-
-                // Save the user locally and in app state
-                currentUser.saveToLocalStorage();
-                setUser(currentUser);
-
-                // Redirect to the main view
-                navigate("/");
-            }
-        };
-
-        initializeUser();
-    }, [isAuthenticated, auth0User, setUser, navigate]);
+        if (currentUser) {
+            navigate("/"); // Redirect when user is initialized
+        }
+    }, [currentUser, navigate]);
 
     return <FullPageLoader message="Initializing your account..." />;
 };
 
-export default LoadingPage;
+export default AuthLandingPage;
